♻️ refactor: rss/oAuth/user/feed/board 서비스 순수 함수 util 분리
