+++ server/src/user/util/createHashedPassword.ts
@@ -1,0 +1,7 @@
+import * as bcrypt from 'bcrypt';
+
+import { SALT_ROUNDS } from '@user/constant/user.constants';
+
+export async function createHashedPassword(password: string) {
+  return await bcrypt.hash(password, SALT_ROUNDS);
+}
