+++ server/test/feed/unit/feed.service.unit.spec.ts
@@ -499,41 +499,4 @@ describe(`${FeedService.name} Unit Test`, () => {
       expect(feedRepository.delete).not.toHaveBeenCalled();
     });
   });
-
-  // getIp는 헤더 문자열 파싱만 하는 순수 로직이라 e2e 대신 단위로 격리 검증한다.
-  // (실제 헤더가 Express→서비스로 흐르는지는 up-view-count e2e가 Redis 상태로 검증)
-  describe('getIp (private)', () => {
-    const getIp = (request: unknown) =>
-      (feedService as unknown as { getIp(request: unknown): string }).getIp(
-        request,
-      );
-
-    const createRequest = (
-      xff: string | string[] | undefined,
-      remoteAddress = '127.0.0.1',
-    ) => ({
-      headers: { 'x-forwarded-for': xff },
-      socket: { remoteAddress },
-    });
-
-    it('x-forwarded-for 단일 IP면 해당 IP를 반환한다.', () => {
-      expect(getIp(createRequest('203.0.113.1'))).toBe('203.0.113.1');
-    });
-
-    it('x-forwarded-for에 여러 IP가 있으면 첫 번째 IP를 trim해 반환한다.', () => {
-      expect(getIp(createRequest('203.0.113.1, 198.51.100.1'))).toBe(
-        '203.0.113.1',
-      );
-    });
-
-    it('x-forwarded-for가 없으면 socket.remoteAddress를 반환한다.', () => {
-      expect(getIp(createRequest(undefined, '10.0.0.5'))).toBe('10.0.0.5');
-    });
-
-    it('x-forwarded-for가 배열(문자열 아님)이면 socket.remoteAddress로 폴백한다.', () => {
-      expect(getIp(createRequest(['203.0.113.1'], '10.0.0.5'))).toBe(
-        '10.0.0.5',
-      );
-    });
-  });
 });
